+++ src/all/kavita/src/eu/kanade/tachiyomi/extension/all/kavita/dto/MetadataDto.kt
@@ -63,6 +63,9 @@ data class MetadataPayload(
     var language: ArrayList<String> = arrayListOf<String>(),
     var libraries: ArrayList<Int> = arrayListOf<Int>(),
     var pubStatus: ArrayList<Int> = arrayListOf<Int>(),
+    var seriesNameQuery: String = "",
+    var releaseYearRangeMin: Int = 0,
+    var releaseYearRangeMax: Int = 0,
 
     var peopleWriters: ArrayList<Int> = arrayListOf<Int>(),
     var peoplePenciller: ArrayList<Int> = arrayListOf<Int>(),
@@ -74,5 +77,4 @@ data class MetadataPayload(
     var peoplePublisher: ArrayList<Int> = arrayListOf<Int>(),
     var peopleCharacter: ArrayList<Int> = arrayListOf<Int>(),
     var peopleTranslator: ArrayList<Int> = arrayListOf<Int>(),
-
 )